feat(bs): add inv save/get if
